스크립트 파일 정리
