+++ script/jsonData.js
@@ -1,0 +1,9 @@
+class JSONData {
+    constructor(type, value, child) {
+        this.type = type
+        this.value = value
+        this.child = child
+    }
+}
+
+module.exports = JSONData;
